+++ src/api.rs
@@ -445,7 +445,7 @@ mod tests {
     use crate::model::trader::order::ExecutionType;
     use crate::model::trader::order_request::InstrumentRequest;
     use crate::model::trader::preview_order::Instruction;
-    use crate::token::local_server::LocalServerMessenger;
+    use crate::token::channel_messenger::local_server::LocalServerMessenger;
     use crate::token::TokenChecker;
 
     async fn client() -> Api<TokenChecker<LocalServerMessenger>> {
